Update AWS broker catalog for RDS instances
This changeset updates our AWS broker catalog to account for the following:

- We are not currently charging for any RDS instances
- We have added redundant versions of our new micro-psql and small-mysql plans

Our pricing will change again in FY21, but until that time there is no reason to display pricing information as we are not currently charging for the instances themselves (only extra storage above 1TB).  The existing pricing information is flowing through to the customer dashboard, which causes confusion about our billing.  Setting this to 0 should help alleviate that disrepancy.

Note that the new redundant plans (and all existing plans) are not set to free; free plans are available in sandboxes, and only the non-redundant micro-psql and small-mysql plans (as well as the shared versions for the time being) should be available for sandboxes.