Update RDS instances for AWS broker
This changeset is a part of multiple steps that are intended to update our AWS broker RDS instances to more current versions for their respective DB engines and reduce the instance class sizes of the dev and staging environment instances.  The steps are as follows:

- Update the DB versions for dev and staging to latest available versions (this commit)
- Update the instance class sizes for dev and staging
- Update the DB versions for production to latest available

These changes also introduce a few improvements to bring more parity between the management of the MySQL and PostgreSQL instances, particularly around the parameter group family names.