Add full support for specifying a version with RDS instances
This changeset adds new functionality to fully support creating and modifying RDS instances with specific versions for PostgreSQL and MySQL instances.  A user can specify a custom parameter to provide a version and as long as it is a supported version in AWS GovCloud (and on a valid upgrade path in the case of moodification), that version will be used.

- Adds support to create new MySQL instances with a specific version
- Adds support to modify existing MySQL and PostgreSQL instances with a specific version
- Adds minimum and maximum supported version numbers to the plan catalog
- Fixes an issue with the BackupRetentionPeriod catalog property; fixes associated plan information to account for this
- New tests added to cover this functionality
- New smoke tests to cover the full breadth of version support
- Pipeline cleanup