Fixes the provisioning of new database instances
AWS recently changed the instance class offerings depending on the version of database you selected, and in the case of PostgreSQL the default is now version 13 as well.  Until we build in more support and update our plans, we need to make sure customers can provision the databases that we do currently support.  This changeset ensures that a default version is provided so that the instance class offerings are still valid.